+++ HealthCheck/HealthCheckAPI/CHANGELOG.md
@@ -21,4 +21,6 @@ The following steps occurs after project was generated for our purpose:
 - No longer need to add Nuget package Microsoft.AspNetcore.Diagnostics.HealthChecks
 - Update `Program.cs` add HealthCheck to the service object
 - Update `Program.cs` add UseHealthCheck and add /api/health as a param as type PathString to main WebApplication object
+- Add new folder `./HealthCheck` and new file ICMPHealthCheck.cs
+- Update `Program.cs` to HealthChecks() add AddCheck<ICMPHealthCheck>("ICMP")
 -